moved ValveAuthException to higher namespace
