+++ df-valve/src/main/java/com/hashmapinc/tempus/witsml/valve/ValveAuthException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.tempus.witsml.valve.dot;
+package com.hashmapinc.tempus.witsml.valve;
 
 public class ValveAuthException extends Exception {
