+++ df-valve/src/main/java/com/hashmapinc/tempus/witsml/valve/dot/DotClient.java
@@ -23,6 +23,7 @@ import java.util.logging.Logger;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.interfaces.DecodedJWT;
+import com.hashmapinc.tempus.witsml.valve.ValveAuthException;
 import com.hashmapinc.tempus.witsml.valve.ValveException;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
